Added bitstamp & bitfinex

+++ exchange/exchange.go
@@ -1,5 +1,5 @@
 package exchange
 
 type Exchange interface {
-	LatestPrice(ticker string) float32
+	LatestPrice(ticker string) float64
 }
